refactor(30-01): rewrite 58 consumer imports to canonical currency, delete wrapper
Change all imports from #lib/formatters/currency to #lib/utils/currency.
Update settings-page.test.tsx mock target. Delete the re-export wrapper
file, eliminating the indirection layer per Zero Tolerance Rule #2.
