+++ src/app/(owner)/financials/cash-flow/cash-flow-breakdown.tsx
@@ -1,6 +1,6 @@
 import { ArrowUpCircle, ArrowDownCircle } from 'lucide-react'
 import { BlurFade } from '#components/ui/blur-fade'
-import { formatCents } from '#lib/formatters/currency'
+import { formatCents } from '#lib/utils/currency'
 
 interface FlowItem {
 	category: string
