+++ src/app/(owner)/financials/expenses/_components/expense-table.tsx
@@ -26,7 +26,7 @@ import {
 	TableRow
 } from '#components/ui/table'
 import { BlurFade } from '#components/ui/blur-fade'
-import { formatCents } from '#lib/formatters/currency'
+import { formatCents } from '#lib/utils/currency'
 import type { Expense } from '#hooks/api/use-expense-mutations'
 import { getCategoryBadge, EXPENSE_CATEGORIES } from './expense-category-badge'
