+++ src/app/(owner)/financials/financials-quick-links.tsx
@@ -11,7 +11,7 @@ import {
 	CreditCard
 } from 'lucide-react'
 import { BlurFade } from '#components/ui/blur-fade'
-import { formatCents } from '#lib/formatters/currency'
+import { formatCents } from '#lib/utils/currency'
 
 interface QuickLinkCardProps {
 	href: string
