+++ src/app/(owner)/financials/income-statement/income-statement-page-breakdowns.tsx
@@ -1,6 +1,6 @@
 import { ArrowUpRight, ArrowDownRight } from 'lucide-react'
 import { BlurFade } from '#components/ui/blur-fade'
-import { formatCents } from '#lib/formatters/currency'
+import { formatCents } from '#lib/utils/currency'
 
 interface BreakdownItem {
 	label: string
