+++ src/app/(owner)/financials/income-statement/income-statement-page-net-summary.tsx
@@ -1,5 +1,5 @@
 import { BlurFade } from '#components/ui/blur-fade'
-import { formatCents } from '#lib/formatters/currency'
+import { formatCents } from '#lib/utils/currency'
 
 interface PreviousPeriod {
 	netIncome: number
