+++ src/app/(owner)/financials/payouts/_components/payout-history-table.tsx
@@ -8,7 +8,7 @@ import { DataTableToolbar } from '#components/data-table/data-table-toolbar'
 import { useDataTable } from '#hooks/use-data-table'
 import type { ColumnDef } from '@tanstack/react-table'
 import type { Payout } from '#hooks/api/use-stripe-connect'
-import { formatCurrency } from '#lib/formatters/currency'
+import { formatCurrency } from '#lib/utils/currency'
 import { formatDate } from '#lib/formatters/date'
 import { getPayoutStatusBadge } from './payout-status-badge'
