+++ src/app/(tenant)/tenant/lease/page.tsx
@@ -18,7 +18,7 @@ import { CardLayout } from '#components/ui/card-layout'
 import { Skeleton } from '#components/ui/skeleton'
 import { useQuery } from '@tanstack/react-query'
 import { leaseQueries } from '#hooks/api/query-keys/lease-keys'
-import { formatCurrency } from '#lib/formatters/currency'
+import { formatCurrency } from '#lib/utils/currency'
 import { formatDate } from '#lib/formatters/date'
 import { Calendar, DollarSign, FileText, Home, MapPin } from 'lucide-react'
 import Link from 'next/link'
