+++ src/components/leases/detail/lease-details.client.tsx
@@ -18,7 +18,7 @@ import {
 
 import { LeaseDetailsSkeleton } from './lease-details-skeleton'
 import { LeaseHeader } from './lease-header'
-import { formatCurrency } from '#lib/formatters/currency'
+import { formatCurrency } from '#lib/utils/currency'
 import { getOrdinalSuffix } from '#lib/formatters/date'
 import { generateTimelineEvents } from './lease-detail-utils'
 import { LeaseDetailsTab } from './lease-details-tab'
