+++ src/components/payments/rent-collection/overdue-payments-alert.tsx
@@ -20,7 +20,7 @@ import {
 	TableHeader,
 	TableRow
 } from '#components/ui/table'
-import { formatCents } from '#lib/formatters/currency'
+import { formatCents } from '#lib/utils/currency'
 
 interface OverduePayment {
 	id: string
