+++ src/components/payments/rent-collection/payment-history-tab.tsx
@@ -8,7 +8,7 @@ import {
 	TableHeader,
 	TableRow
 } from '#components/ui/table'
-import { formatCents } from '#lib/formatters/currency'
+import { formatCents } from '#lib/utils/currency'
 import { getPaymentStatusBadge } from './status-utils'
 
 interface PaymentHistoryItem {
