+++ src/components/tenant-portal/tenant-portal.tsx
@@ -5,7 +5,7 @@ import Link from 'next/link'
 import { BlurFade } from '#components/ui/blur-fade'
 import { BorderBeam } from '#components/ui/border-beam'
 import { Button } from '#components/ui/button'
-import { formatCurrency } from '#lib/formatters/currency'
+import { formatCurrency } from '#lib/utils/currency'
 import { formatDate } from '#lib/formatters/date'
 import { TenantStatsCards, type RentStatus } from './tenant-stats-cards'
 import { PaymentHistoryCard } from './payment-history-card'
